+++ server/src/game/mode/modeManager.ts
@@ -7,7 +7,7 @@ import { GameItem } from "common/scripts/definitions/game_defs.ts";
 import { type Group, type Team } from "./teams.ts";
 import { LevelEnemys } from "common/scripts/config/level_definition.ts";
 import { JoinPacket } from "common/scripts/packets/join_packet.ts";
-import { LootAditional } from "common/scripts/others/constants.ts";
+import { LootAditional, LootSetting } from "common/scripts/others/constants.ts";
 
 export interface GameRules{
     humans:{
@@ -75,7 +75,7 @@ export interface GameRules{
     leader:{
         kills_min:number
     }
-    loot_settings:LootAditional
+    loot_settings:LootSetting
 }
 export abstract class ModeManager{
     game!:Game